Refactored and separated into files
